requirements.txt
langchain
langchain-openai
langchain-ollama
langchain-google-genai
langchain-anthropic
pydantic
pillow